requirements.txt
grpcio-tools==1.70.0
grpcio>=0.0
protobuf>0.0
python_dotenv>0.0
kafka-python>=0.0